backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
databases==0.8.0
environs==10.3.0
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi_cors==0.0.6
h11==0.14.0
idna==3.6
marshmallow==3.20.2
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
